名前空間を意識したルーティングをする -> APIのバージョン管理が容易になるため
作業手順
- routerを変更
- controllerのディレクトリ構成を変更
- postmanで動作確認
- react側でのpathを変更
