+++ api/app/controllers/api/v1/ip_addresses_controller.rb
@@ -1,0 +1,10 @@
+module Api
+	module V1
+		class IpAddressesController < ApplicationController
+			def get
+				@res = { ip: request.remote_ip }
+				render json: { status: 'SUCCESS', data: @res }
+			end
+		end
+	end
+end
