--- api/app/controllers/ip_addresses_controller.rb
@@ -1,6 +1,0 @@
-class IpAddressesController < ApplicationController
-	def get
-		@res = { ip: request.remote_ip }
-		render json: { status: 'SUCCESS', data: @res }
-	end
-end
